refactor: widgets folder
