+++ src/components/widgets/weather/DialogCreateWeatherWidget/index.module.css
@@ -1,0 +1,6 @@
+.weatherOptionsBlock {
+    display: flex;
+    gap: 8px;
+    align-items: center;
+    justify-content: start;
+}
