+++ src/components/widgets/weather/WeatherWidget/index.module.css
@@ -1,3 +1,25 @@
+.widget {
+    flex: 0 0 46px;
+    border-radius: 15px;
+    width: 44px;
+    height: 44px;
+    background-color: var(--border-color);
+
+    align-items: flex-start;
+    justify-content: center;
+    margin: 0px auto;
+    transition: 0.3s ease-in-out;
+}
+
+.widget img {
+    border-radius: 50%;
+}
+
+.widget:hover {
+    background-color: antiquewhite;
+}
+
+
 .weatherIcon {
     font-size: 1.4rem;
     margin: 0;
@@ -39,13 +61,6 @@
     line-height: 1.2;
 }
 
-.weatherOptionsBlock {
-    display: flex;
-    gap: 8px;
-    align-items: center;
-    justify-content: start;
-}
-
 @media (max-width: 650px) {
     .weatherIcon {
         font-size: 1.4rem;
@@ -85,40 +100,3 @@
         display: none;
     }
 }
-
-
-
-/* .widget {
-    display: flex;
-    width: 200px;
-    height: 50px;
-    gap: 12px;
-    justify-content: center;
-    align-items: center;
-    background-color: var(--color-bg-card);
-    top: 20px;
-    right: 20px;
-    transition: 0.2s ease-in-out;
-    border-radius: 15px;
-    padding: 12px 16px;
-    box-shadow: 0 2px 10px rgba(0, 0, 0, 0.1);
-  }
-  
-  .widget:hover {
-    transform: scale(1.02);
-    box-shadow: 0 4px 20px rgba(0, 0, 0, 0.15);
-  }
-  
-  @media (max-width: 650px) {
-    .widget {
-      padding: 10px 15px;
-      width: 160px;
-      height: 50px;
-    }
-  }
-  
-  @media (max-width: 450px) {
-    .widget {
-      width: 70px;
-    }
-  } */
